refactor: extract callbacks package
