+++ api/chatbot/callbacks/conversation.py
@@ -1,0 +1,26 @@
+from typing import Any, Optional
+from uuid import UUID
+
+from langchain.callbacks.base import AsyncCallbackHandler
+
+from chatbot.models import Conversation
+from chatbot.utils import utcnow
+
+
+class UpdateConversationCallbackHandler(AsyncCallbackHandler):
+    def __init__(self, conversation_id: str):
+        self.conversation_id: str = conversation_id
+
+    async def on_chain_end(
+        self,
+        outputs: dict[str, Any],
+        *,
+        run_id: UUID,
+        parent_run_id: Optional[UUID] = None,
+        tags: Optional[list[str]] = None,
+        **kwargs: Any,
+    ) -> None:
+        """Run when chain ends running."""
+        conv = await Conversation.get(self.conversation_id)
+        conv.updated_at = utcnow()
+        await conv.save()
